backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coverage==7.6.1
cryptography==44.0.3
Django==4.2.20
django-appconf==1.0.6
django-cors-headers==4.3.1
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
exceptiongroup==1.2.2
idna==3.10
importlib-resources==6.4.5
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==25.0
pkgutil-resolve-name==1.3.10
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==7.4.4
pytest-django==4.5.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
sendgrid==6.10.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==7.14.0
tomli==2.2.1
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
json_log_formatter==0.5.1
psycopg2-binary==2.9.9
